test(tg_client): split tests
